add double support to Str() (#581)

+++ src/stdlib/String.ts
@@ -144,11 +144,11 @@ export const Mid = new Callable(
 );
 
 /**
- * Return a string from a float. If it is positive, prefix it with a space.
+ * Return a string from a float/double. If it is positive, prefix it with a space.
  */
 export const Str = new Callable("Str", {
     signature: {
-        args: [new StdlibArgument("value", ValueKind.Float)],
+        args: [new StdlibArgument("value", ValueKind.Double)],
         returns: ValueKind.String,
     },
     impl: (interpreter: Interpreter, value: Float): BrsString => {
